Use engine to flush records of model
Instead of iterating over ids from the database we use the engine itself to flush the records of an index/model.

This solves the problem where models would get out of sync with the search records and old records weren't flushed. With this implementation all records are always removed.

If there are engines which don't provide this functionality they can still iterate over the model ids if they want to.

See https://github.com/laravel/scout/issues/101